+++ src/Libraries/SmartStore.Core/Domain/Forums/ForumSettings.cs
@@ -11,23 +11,24 @@ namespace SmartStore.Core.Domain.Forums
 			TopicSubjectMaxLength = 450;
 			PostMaxLength = 4000;
 			StrippedTopicMaxLength = 45;
-			TopicsPageSize = 10;
-			PostsPageSize = 10;
-            SearchResultsPageSize = 12;
+			TopicsPageSize = 20;
+			PostsPageSize = 20;
+            ForumFeedCount = 20;
+            SearchResultsPageSize = 20;
+            LatestCustomerPostsPageSize = 20;
             AllowSorting = true;
-            LatestCustomerPostsPageSize = 10;
-			ShowCustomersPostCount = true;
+            ShowCustomersPostCount = true;
 			ForumEditor = EditorType.BBCodeEditor;
 			SignaturesEnabled = true;
-			PrivateMessagesPageSize = 10;
-			ForumSubscriptionsPageSize = 10;
+			PrivateMessagesPageSize = 20;
+			ForumSubscriptionsPageSize = 20;
 			PMSubjectMaxLength = 450;
 			PMTextMaxLength = 4000;
 			HomePageActiveDiscussionsTopicCount = 5;
 			ActiveDiscussionsPageTopicCount = 50;
 			ActiveDiscussionsFeedCount = 25;
-			ForumFeedCount = 10;
-		}
+            AllowCustomersToVoteOnPosts = true;
+        }
 		
 		/// <summary>
         /// Gets or sets a value indicating whether forums are enabled
@@ -63,6 +64,16 @@ namespace SmartStore.Core.Domain.Forums
         /// Gets or sets a value indicating whether customers are allowed to delete posts that they created
         /// </summary>
         public bool AllowCustomersToDeletePosts { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether customer can vote on posts
+        /// </summary>
+        public bool AllowCustomersToVoteOnPosts { get; set; }
+
+        /// <summary>
+        /// Gets or sets a value indicating whether guests are allowed to vote on posts
+        /// </summary>
+        public bool AllowGuestsToVoteOnPosts { get; set; }
 
         /// <summary>
         /// Gets or sets maximum length of topic subject